+++ CHANGELOG.md
@@ -4,6 +4,14 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- indexes added to stream_id and position_in_stream in event_log table 
+### Changed
+- DefaultEventStoreDataSourceProvider changed to a singleton, 
+so the caching of the DataSource works properly
+- Test util class DatabaseCleaner has an additional method 
+'cleanEventStoreTables(...)' for truncating specified tables 
+in the event-store
 
 ## [7.0.5] - 2020-07-08
 ### Changed